refactor: rename sample_types.js to sample_type.js for consistency
