+++ acl-js/src-js/sample.js
@@ -22,7 +22,7 @@
 // SOFTWARE.
 ////////////////////////////////////////////////////////////////////////////////
 
-import { SampleType, isSampleType } from './sample_types.js'
+import { SampleType, isSampleType } from './sample_type.js'
 import { QVV } from './qvv.js'
 
 ////////////////////////////////////////////////////////////////////////////////
